requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
setuptools==69.5.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==2.3.7
WTForms==3.1.2
